Data/Tools/requirements.txt
regex
PyPDF2<3.0
camelot-py[base]
opencv-python
